Move macos-specific functions to macos module in mullvad-daemon
