+++ mullvad-daemon/src/lib.rs
@@ -11,10 +11,10 @@ mod cleanup;
 pub mod device;
 mod dns;
 pub mod exception_logging;
-#[cfg(target_os = "macos")]
-pub mod exclusion_gid;
 mod geoip;
 pub mod logging;
+#[cfg(target_os = "macos")]
+mod macos;
 #[cfg(not(target_os = "android"))]
 pub mod management_interface;
 mod migrations;
@@ -544,8 +544,8 @@ where
     ) -> Result<Self, Error> {
         #[cfg(target_os = "macos")]
         let exclusion_gid = {
-            bump_filehandle_limit();
-            exclusion_gid::set_exclusion_gid().map_err(Error::GroupIdError)?
+            macos::bump_filehandle_limit();
+            macos::set_exclusion_gid().map_err(Error::GroupIdError)?
         };
 
         mullvad_api::proxy::ApiConnectionMode::try_delete_cache(&cache_dir).await;
@@ -2174,42 +2174,5 @@ fn new_selector_config(settings: &Settings) -> SelectorConfig {
         bridge_state: settings.get_bridge_state(),
         bridge_settings: settings.bridge_settings.clone(),
         obfuscation_settings: settings.obfuscation_settings.clone(),
-    }
-}
-
-/// Bump filehandle limit
-#[cfg(target_os = "macos")]
-pub fn bump_filehandle_limit() {
-    let mut limits = libc::rlimit {
-        rlim_cur: 0,
-        rlim_max: 0,
-    };
-    // SAFETY: `&mut limits` is a valid pointer parameter for the getrlimit syscall
-    let status = unsafe { libc::getrlimit(libc::RLIMIT_NOFILE, &mut limits) };
-    if status != 0 {
-        log::error!(
-            "Failed to get file handle limits: {}-{}",
-            io::Error::from_raw_os_error(status),
-            status
-        );
-        return;
-    }
-
-    const INCREASED_FILEHANDLE_LIMIT: u64 = 1024;
-    // if file handle limit is already big enough, there's no reason to decrease it.
-    if limits.rlim_cur >= INCREASED_FILEHANDLE_LIMIT {
-        return;
-    }
-
-    limits.rlim_cur = INCREASED_FILEHANDLE_LIMIT;
-    // SAFETY: `&limits` is a valid pointer parameter for the getrlimit syscall
-    let status = unsafe { libc::setrlimit(libc::RLIMIT_NOFILE, &limits) };
-    if status != 0 {
-        log::error!(
-            "Failed to set file handle limit to {}: {}-{}",
-            INCREASED_FILEHANDLE_LIMIT,
-            io::Error::from_raw_os_error(status),
-            status
-        );
     }
 }
